+++ app/monitoring/update-delivery-from-response.js
@@ -1,4 +1,4 @@
-const { DELIVERED, PERMANENT_FAILURE, TEMPORARY_FAILURE, TECHNICAL_FAILURE } = require('../delivery-statuses')
+const { DELIVERED, PERMANENT_FAILURE, TEMPORARY_FAILURE, TECHNICAL_FAILURE } = require('../statuses')
 const completeDelivery = require('./complete-delivery')
 const createFailure = require('./create-failure')
 const rescheduleDelivery = require('./reschedule-delivery')
